changed data fetching model to ssr instead of api routes where needed
