--- services/employee.service.ts
@@ -1,10 +1,0 @@
-'use client'
-
-
-const getEmployeeShifts = async (employeeId: Array<string>) => {
-
-}
-
-export const employeeService = {
-    getEmployeeShifts
-}
